// src/zigbee/types.ts
type State = '' | 'ON' | 'OFF' | 'TOGGLE' | 'LOCK' | 'UNLOCK';
export type SystemMode = 'off' | 'heat' | 'cool' | 'auto';

export interface ZigBeeControllerConfig {
  port?: string;
  panId?: number;
  channels: number[];
  databasePath: string;
  adapter: 'zstack' | 'deconz' | 'zigate';
}

export type RunningState = 'idle' | 'heat' | 'cool';

export type ButtonAction =
  | 'on'
  | 'off'
  | 'toggle'
  | 'toggle_hold'
  | 'toggle_release'
  | 'arrow_left_click'
  | 'arrow_left_hold'
  | 'arrow_left_release'
  | 'arrow_right_click'
  | 'arrow_right_hold'
  | 'arrow_right_release'
  | 'brightness_down'
  | 'brightness_up'
  | 'brightness_up_click'
  | 'brightness_up_hold'
  | 'brightness_up_release'
  | 'brightness_down_click'
  | 'brightness_down_hold'
  | 'brightness_down_release'
  | 'brightness_move_down'
  | 'brightness_move_up'
  | 'brightness_stop'
  | 'vibration'
  | 'tilt'
  | 'button_1_hold'
  | 'button_2_hold'
  | 'button_3_hold'
  | 'button_4_hold'
  | 'button_5_hold'
  | 'button_6_hold'
  | 'button_1_release'
  | 'button_2_release'
  | 'button_3_release'
  | 'button_4_release'
  | 'button_5_release'
  | 'button_6_release'
  | 'button_1_single'
  | 'button_2_single'
  | 'button_3_single'
  | 'button_4_single'
  | 'button_5_single'
  | 'button_6_single'
  | 'button_1_double'
  | 'button_2_double'
  | 'button_3_double'
  | 'button_4_double'
  | 'button_5_double'
  | 'button_6_double'
  | 'button_1_click'
  | 'button_2_click'
  | 'button_3_click'
  | 'button_4_click'
  | 'button_5_click'
  | 'button_6_click'
  | 'drop'
  | 'single'
  | 'hold'
  | 'double'
  | 'play_pause'
  | 'release'
  | 'single_left'
  | 'single_right'
  | 'single_both'
  | 'double_left'
  | 'double_right'
  | 'double_both'
  | 'hold_left'
  | 'hold_right'
  | 'hold_both';

export type ClickAction =
  | 'on'
  | 'off'
  | 'brightness_down'
  | 'brightness_up'
  | 'play_pause'
  | 'single'
  | 'double'
  | 'hold'
  | 'release';

export interface DeviceState {
  state?: State;
  state_left?: State;
  state_right?: State;
  brightness?: number; // Value between 0 and 255
  brightness_percent?: number; // 0-100
  // Color temperature in Reciprocal MegaKelvin, a.k.a. Mirek scale.
  // Mirek = 1,000,000 / Color Temperature in Kelvin
  // Values typically between 50 and 400. The higher the value, the warmer the color.
  color_temp?: number;
  color?: {
    // RGB color
    hue?: number;
    s?: number;
    x?: number;
    y?: number;

    rgb?: string;
  };
  // Blinks the bulbs, possible values:
  // - "select": single blink
  // - "lselect": blinking for a longer time
  // - "none": stop blinking
  alert?: 'select' | 'lselect' | 'none';
  temperature?: number;
  local_temperature?: number;
  current_heating_setpoint?: number;
  system_mode?: SystemMode;
  running_state?: RunningState;
  humidity?: number;
  click?: ClickAction;
  state_l1?: 'ON' | 'OFF' | 'TOGGLE';
  state_l2?: 'ON' | 'OFF' | 'TOGGLE';
  lock_state?: '' | 'not_fully_locked' | 'locked' | 'unlocked';
  battery?: number;
  current?: number;
  power?: number;
  voltage?: number;
  illuminance?: number;
  illuminance_lux?: number;
  contact?: boolean;
  action?: ButtonAction;
  occupancy?: boolean;
  tamper?: boolean;
  battery_low?: boolean;
  strength?: number;
  linkquality?: number;
  sensitivity?: 'low' | 'medium' | 'high';
  water_leak?: boolean;
  gas?: boolean;
  smoke?: boolean;
  // Philips HUE specifics
  hue_power_on_behavior?: 'default' | 'on' | 'off' | 'recover';
  hue_power_on_brightness?: number;
  hue_power_on_color_temperature?: number;
  hue_power_on_color?: string;
  // Tuya specific
  position?: number;
}